Updated html and css files
Added media queries to allow for columns to wrap once at smartphone size and rearanged the colums from their original pattern of having the pictures and descriptions alternate the side of the page they're on.  Changed to just having all pictures on the left, and all descriptions on the right.